move UserToPostRelTypes to /toPost
